move utilities.py to utils/utilities.py
